refactor: userService 추상화
