+++ app/src/main/java/com/techcourse/service/TxUserService.java
@@ -1,0 +1,61 @@
+package com.techcourse.service;
+
+import com.interface21.dao.DataAccessException;
+import com.interface21.jdbc.datasource.DataSourceUtils;
+import com.interface21.transaction.support.TransactionSynchronizationManager;
+import com.techcourse.domain.User;
+import java.sql.Connection;
+import java.sql.SQLException;
+import javax.sql.DataSource;
+
+public class TxUserService implements UserService {
+
+    private final UserService userService;
+    private final DataSource dataSource;
+
+    public TxUserService(final UserService userService, final DataSource dataSource) {
+        this.userService = userService;
+        this.dataSource = dataSource;
+    }
+
+    @Override
+    public User findById(final long id) {
+        return userService.findById(id);
+    }
+
+    @Override
+    public void save(final User user) {
+        userService.save(user);
+    }
+
+    @Override
+    public void changePassword(final long id, final String newPassword, final String createdBy) {
+        Connection connection = null;
+        try {
+            connection = DataSourceUtils.getConnection(dataSource);
+            connection.setAutoCommit(false);
+            TransactionSynchronizationManager.bindResource(dataSource, connection);
+
+            userService.changePassword(id, newPassword, createdBy);
+
+            connection.commit();
+        } catch (SQLException e) {
+            rollbackTransaction(connection);
+            throw new DataAccessException(e);
+        } finally {
+            DataSourceUtils.releaseConnection(connection, dataSource);
+            TransactionSynchronizationManager.unbindResource(dataSource);
+        }
+    }
+
+    private void rollbackTransaction(final Connection connection) {
+        if (connection != null) {
+            try {
+                connection.setAutoCommit(true);
+                connection.rollback();
+            } catch (SQLException e) {
+                throw new DataAccessException(e);
+            }
+        }
+    }
+}
